+++ projboard/models/article.py
@@ -1,56 +1,8 @@
 from django.db import models
 from django.utils import timezone
 from django.core.exceptions import ObjectDoesNotExist
-
-
-class Subject(models.Model):
-    """
-    Subject model
-
-    name- A string, the subject name
-    """
-    name = models.CharField(max_length=100)
-
-    @staticmethod
-    def get_subject_by_name(name):
-        """
-        Method get subject name and return Subject object if exist, otherwise None
-        :param name: the name to search
-        :return: Subject object
-        """
-        try:
-            subject = User.objects.get(name=name)
-        except ObjectDoesNotExist:
-            return None
-        return subject
-
-
-class User(models.Model):
-    """
-    User model
-
-    email- User's Email
-    password- User's password
-    name- User's name
-    nickname- User's nickname
-    """
-    email = models.EmailField(max_length=150, unique=True)
-    password = models.CharField(max_length=70)
-    name = models.CharField(max_length=100)
-    nickname = models.CharField(max_length=100, unique=True)
-
-    @staticmethod
-    def get_user_by_nickname(nickname):
-        """
-        Method get nickname and return User object if exist, otherwise None
-        :param nickname: the nickname to search
-        :return: User object
-        """
-        try:
-            user = User.objects.get(nickname=nickname)
-        except ObjectDoesNotExist:
-            return None
-        return user
+from .subject import Subject
+from .user import User
 
 
 class Article(models.Model):
