+++ projboard/models/user.py
@@ -1,0 +1,30 @@
+from django.db import models
+from django.core.exceptions import ObjectDoesNotExist
+
+
+class User(models.Model):
+    """
+    User model
+
+    email- User's Email
+    password- User's password
+    name- User's name
+    nickname- User's nickname
+    """
+    email = models.EmailField(max_length=150, unique=True)
+    password = models.CharField(max_length=70)
+    name = models.CharField(max_length=100)
+    nickname = models.CharField(max_length=100, unique=True)
+
+    @staticmethod
+    def get_user_by_nickname(nickname):
+        """
+        Method get nickname and return User object if exist, otherwise None
+        :param nickname: the nickname to search
+        :return: User object
+        """
+        try:
+            user = User.objects.get(nickname=nickname)
+        except ObjectDoesNotExist:
+            return None
+        return user
